added visual loading and error for ActorPage section with timing of 1 second
